aiserver/requirements.txt
fastapi
uvicorn
pymysql
pandas
python-dotenv
numpy
python-multipart==0.0.10
torch==2.4.1
torchvision==0.19.1
Pillow==10.3.0